More int -> pointer fixes

+++ src/age/graph/vertex_arrays.h
@@ -424,7 +424,7 @@ public:
 
   varray_index(void)
   {
-    handle = ERROR;
+    handle = (GLuint)ERROR;
     size = 0;
     flag = 0;
   }
